# email-service/proto/api/email/v1/email_service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/email/v1/email_service.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "EmailService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/email": {
      "get": {
        "operationId": "EmailService_ListEmails",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListEmailsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageToken",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "EmailService"
        ]
      }
    },
    "/api/v1/email/send": {
      "post": {
        "operationId": "EmailService_SendEmail",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1SendEmailResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1SendEmailRequest"
            }
          }
        ],
        "tags": [
          "EmailService"
        ]
      }
    },
    "/api/v1/email/{id}/status": {
      "get": {
        "operationId": "EmailService_GetEmailStatus",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetEmailStatusResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "EmailService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1Email": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "to": {
          "type": "string"
        },
        "subject": {
          "type": "string"
        },
        "body": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "sentAt": {
          "type": "string"
        }
      },
      "required": [
        "to",
        "subject",
        "body"
      ]
    },
    "v1GetEmailStatusResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "sentAt": {
          "type": "string"
        }
      }
    },
    "v1ListEmailsResponse": {
      "type": "object",
      "properties": {
        "emails": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Email"
          }
        },
        "nextPageToken": {
          "type": "string"
        }
      }
    },
    "v1SendEmailRequest": {
      "type": "object",
      "properties": {
        "to": {
          "type": "string"
        },
        "subject": {
          "type": "string"
        },
        "body": {
          "type": "string"
        }
      },
      "required": [
        "to",
        "subject",
        "body"
      ]
    },
    "v1SendEmailResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      }
    }
  }
}